+++ src/OTSubscriber.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { isNull, isUndefined, each, isEqual, isEmpty } from 'underscore';
 import { OT, nativeEvents, setNativeEvents, removeNativeEvents } from './OT';
 import OTSubscriberView from './views/OTSubscriberView';
-import { sanitizeSubscriberEvents, sanitizeProperties, sanitizeFrameRate, sanitizeResolution } from './helpers/OTSubscriberHelper';
+import { sanitizeSubscriberEvents, sanitizeProperties, sanitizeFrameRate, sanitizeResolution, sanitizeAudioVolume } from './helpers/OTSubscriberHelper';
 import { getOtrnErrorEventHandler, sanitizeBooleanProperty } from './helpers/OTHelper';
 import OTContext from './contexts/OTContext';
 
@@ -18,6 +18,7 @@ export default class OTSubscriber extends Component {
     this.componentEvents = {
       streamDestroyed: Platform.OS === 'android' ? 'session:onStreamDropped' : 'session:streamDestroyed',
       streamCreated: Platform.OS === 'android' ? 'session:onStreamReceived' : 'session:streamCreated',
+      captionReceived: Platform.OS === 'android' ? 'session:onCaptionText' : 'subscriber:subscriberCaptionReceived:',
     };
     this.componentEventsArray = Object.values(this.componentEvents);
     this.otrnEventHandler = getOtrnErrorEventHandler(this.props.eventHandlers);
@@ -40,18 +41,24 @@ export default class OTSubscriber extends Component {
     const { streamProperties } = this.props;
     if (!isEqual(this.state.streamProperties, streamProperties)) {
       each(streamProperties, (individualStreamProperties, streamId) => {
-        const { subscribeToAudio, subscribeToVideo, preferredResolution, preferredFrameRate } = individualStreamProperties;
+        const { subscribeToAudio, subscribeToVideo, subscribeToCaptions, preferredResolution, preferredFrameRate, audioVolume } = individualStreamProperties;
         if (subscribeToAudio !== undefined) {
           OT.subscribeToAudio(streamId, sanitizeBooleanProperty(subscribeToAudio));
         }
         if (subscribeToVideo !== undefined) {
           OT.subscribeToVideo(streamId, sanitizeBooleanProperty(subscribeToVideo));
         }
+        if (subscribeToCaptions !== undefined) {
+          OT.subscribeToCaptions(streamId, sanitizeBooleanProperty(subscribeToCaptions));
+        }
         if (preferredResolution !== undefined) {
           OT.setPreferredResolution(streamId, sanitizeResolution(preferredResolution));
         }
         if (preferredFrameRate !== undefined) {
           OT.setPreferredFrameRate(streamId, sanitizeFrameRate(preferredFrameRate));
+        }
+        if (audioVolume !== undefined) {
+          OT.setAudioVolume(streamId, sanitizeAudioVolume(audioVolume));
         }
       });
       this.setState({ streamProperties });
@@ -68,9 +75,10 @@ export default class OTSubscriber extends Component {
     const { subscribeToSelf } = this.state;
     const { streamProperties, properties } = this.props;
     const { sessionId, sessionInfo } = this.context;
-    const subscriberProperties = isNull(streamProperties[stream.streamId]) ?
-      sanitizeProperties(properties) : sanitizeProperties(streamProperties[stream.streamId]);
-    // Subscribe to streams. If subscribeToSelf is true, subscribe also to his own stream
+    const subscriberProperties = streamProperties[stream.streamId] ?
+      sanitizeProperties(streamProperties[stream.streamId]) :
+      sanitizeProperties(properties);
+      // Subscribe to streams. If subscribeToSelf is true, subscribe also to his own stream
     const sessionInfoConnectionId = sessionInfo && sessionInfo.connection ? sessionInfo.connection.connectionId : null;
     if (subscribeToSelf || (sessionInfoConnectionId !== stream.connectionId)) {
       OT.subscribeToStream(stream.streamId, sessionId, subscriberProperties, (error) => {
@@ -98,6 +106,9 @@ export default class OTSubscriber extends Component {
       }
     });
   }
+  getRtcStatsReport(streamId) {
+    OT.getSubscriberRtcStatsReport(streamId);
+  }
   render() {
     if (!this.props.children) {
       const containerStyle = this.props.containerStyle;
@@ -120,6 +131,7 @@ OTSubscriber.propTypes = {
   eventHandlers: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   streamProperties: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   containerStyle: PropTypes.object, // eslint-disable-line react/forbid-prop-types
+  getRtcStatsReport: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   subscribeToSelf: PropTypes.bool
 };
 
@@ -128,7 +140,9 @@ OTSubscriber.defaultProps = {
   eventHandlers: {},
   streamProperties: {},
   containerStyle: {},
-  subscribeToSelf: false
+  subscribeToSelf: false,
+  getRtcStatsReport: {},
+  subscribeToCaptions: false,
 };
 
 OTSubscriber.contextType = OTContext;