+++ src/helpers/OTPublisherHelper.js
@@ -8,6 +8,8 @@ const sanitizeResolution = (resolution) => {
       return 'MEDIUM';
     case '1280x720':
       return 'HIGH';
+    case '1920x1080':
+      return 'HIGH_1080P';
     default:
       return 'MEDIUM';
   }
@@ -53,6 +55,7 @@ const sanitizeProperties = (properties) => {
       audioTrack: true,
       publishAudio: true,
       publishVideo: true,
+      publishCaptions: false,
       name: '',
       cameraPosition: 'front',
       audioFallbackEnabled: true,
@@ -62,6 +65,7 @@ const sanitizeProperties = (properties) => {
       resolution: sanitizeResolution(),
       videoContentHint: '',
       videoSource: 'camera',
+      scalableScreenshare: false,
     };
   }
   return {
@@ -69,6 +73,7 @@ const sanitizeProperties = (properties) => {
     audioTrack: sanitizeBooleanProperty(properties.audioTrack),
     publishAudio: sanitizeBooleanProperty(properties.publishAudio),
     publishVideo: sanitizeBooleanProperty(properties.publishVideo),
+    publishCaptions: sanitizeBooleanProperty(properties.publishCaptions),
     name: properties.name ? properties.name : '',
     cameraPosition: sanitizeCameraPosition(properties.cameraPosition),
     audioFallbackEnabled: sanitizeBooleanProperty(properties.audioFallbackEnabled),
@@ -78,6 +83,7 @@ const sanitizeProperties = (properties) => {
     resolution: sanitizeResolution(properties.resolution),
     videoContentHint: sanitizeVideoContentHint(properties.videoContentHint),
     videoSource: sanitizeVideoSource(properties.videoSource),
+    scalableScreenshare: Boolean(properties.scalableScreenshare),
   };
 };
 
@@ -91,12 +97,20 @@ const sanitizePublisherEvents = (publisherId, events) => {
       streamDestroyed: 'streamDestroyed',
       error: 'didFailWithError',
       audioLevel: 'audioLevelUpdated',
+      audioNetworkStats: 'audioStats',
+      rtcStatsReport: 'rtcStatsReport',
+      videoNetworkStats: 'videoStats',
+      muteForced: 'muteForced',
     },
     android: {
       streamCreated: 'onStreamCreated',
       streamDestroyed: 'onStreamDestroyed',
       error: 'onError',
       audioLevel: 'onAudioLevelUpdated',
+      audioNetworkStats: 'onAudioStats',
+      rtcStatsReport: 'onRtcStatsReport',
+      videoNetworkStats: 'onVideoStats',
+      muteForced: 'onMuteForced',
     },
   };
   return reassignEvents('publisher', customEvents, events, publisherId);